findItemByLearnTaxonomyTerm
