+++ src/selectors/__tests__/get_tasks_by_explore_taxonomy_term.test.ts
@@ -1,10 +1,9 @@
 // tslint:disable:no-expression-statement no-let
-
-import { findItemByLearnTaxonomyTerm } from '../details/tasks';
 import { ExploreTaxonomyId, TaxonomyTermReference } from '../../stores/taxonomies';
 import { aString } from '../../application/__tests__/helpers/random_test_values';
 import { TaskBuilder } from '../../stores/__tests__/helpers/tasks_helpers';
 import { TaskMap } from '../../stores/tasks';
+import { findItemByLearnTaxonomyTerm } from '../taxonomies/find_item_by_explore_taxonomy_term';
 
 describe('find task by explore taxonomy term', () => {
     let exploreTerm: TaxonomyTermReference;
