+++ src/selectors/details/tasks.ts
@@ -1,31 +1,7 @@
-import { TaxonomyTermReference } from '../../stores/taxonomies';
 import { filterExploreTaxonomyTerms } from '../taxonomies/filter_explore_taxonomy_terms';
 import { ExploreSectionMap, ExploreSection } from '../../stores/explore';
+import { HasTaxonomyTerms } from '../taxonomies/find_item_by_explore_taxonomy_term';
 import * as R from 'ramda';
-
-export interface HasTaxonomyTerms {
-    readonly taxonomyTerms: ReadonlyArray<TaxonomyTermReference>;
-}
-
-export interface MapOfHasTaxonomyTerms<T extends HasTaxonomyTerms> {
-    readonly [property: string]: T;
-}
-
-export function findItemByLearnTaxonomyTerm<T extends HasTaxonomyTerms>(needle: ReadonlyArray<TaxonomyTermReference>,
-    haystack: MapOfHasTaxonomyTerms<T>): ReadonlyArray<T> {
-
-    const needleTerms = filterExploreTaxonomyTerms(needle);
-
-    const matchesTaxonomyTerm = (item: T): boolean => {
-        const haystackTerms = filterExploreTaxonomyTerms(item.taxonomyTerms);
-        const commonTerms = R.intersection(needleTerms, haystackTerms);
-        return R.length(commonTerms) > 0;
-    };
-
-    const find = R.compose(R.values, R.pickBy(matchesTaxonomyTerm));
-
-    return find(haystack);
-}
 
 export function findExploreSectionBy<T extends HasTaxonomyTerms>(needle: T,
     haystack: ExploreSectionMap): ExploreSection | undefined {
